+++ lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'screens/homepage.dart';
-import 'screens/signup.dart';
 
 void main() => runApp(MyApp());
 
@@ -9,7 +8,6 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      debugShowCheckedModeBanner: false,
       title: 'Salve Tempo',
       theme: ThemeData(
         primarySwatch: Colors.blue,
