--- lib/widget/SlideCadastro.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SlideCadastro extends MaterialPageRoute {
-  SlideCadastro({WidgetBuilder builder, RouteSettings settings})
-      : super(builder: builder, settings: settings);
-
-  @override
-  Widget buildTransitions(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child) {
-    Animation<Offset> custom =
-        Tween<Offset>(begin: Offset(0.0, 1.0), end: Offset(0.0, 0.0))
-            .animate(animation);
-    return SlideTransition(position: custom, child: child);
-    //return super.buildTransitions(context, animation, secondaryAnimation, child);
-  }
-}
